+++ docs/snippets/simple_fastapi_app.py
@@ -4,20 +4,25 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI
 
 from grelmicro import Grelmicro
+from grelmicro.coordination import Coordination, LeaderElection
+from grelmicro.coordination.memory import MemoryLeaderElectionBackend
 from grelmicro.log import configure
 from grelmicro.providers.redis import RedisProvider
 from grelmicro.resilience import CircuitBreaker, CircuitBreakers
 from grelmicro.resilience.circuitbreaker.memory import (
     MemoryCircuitBreakerAdapter,
 )
-from grelmicro.sync import LeaderElection, Lock, Sync
+from grelmicro.sync import Lock, Sync
 from grelmicro.task import Tasks
 
 logger = logging.getLogger(__name__)
 
 # === grelmicro ===
 tasks = Tasks()
-leader_election = LeaderElection("leader-election")
+coordination_backend = MemoryLeaderElectionBackend()
+leader_election = LeaderElection(
+    "leader-election", backend=coordination_backend
+)
 tasks.add_task(leader_election)
 
 redis = RedisProvider("redis://localhost:6379/0")
@@ -32,6 +37,7 @@ micro = Grelmicro(
     uses=[
         redis,
         Sync(sync_backend),
+        Coordination(coordination_backend),
         CircuitBreakers(breaker_backend),
         tasks,
     ]